Refactored server utils
