+++ web/src/app/api/ratings/route.ts
@@ -3,7 +3,7 @@ import { getAuthUser } from "@/lib/auth";
 import { parseJsonBody } from "@/lib/validation/parse-body";
 import { createRatingSchema } from "@/lib/validation/schemas/ratings";
 import { rateLimit, getRetryAfterHeader } from "@/lib/rate-limit";
-import { getClientIp } from "@/lib/utils";
+import { getClientIp } from "@/lib/server-utils";
 
 export const dynamic = "force-dynamic";
