+++ web/src/lib/server-utils.ts
@@ -1,0 +1,13 @@
+import { headers } from 'next/headers';
+
+export async function getClientIp(): Promise<string> {
+  const headersList = await headers();
+  const forwardedFor = headersList.get('x-forwarded-for');
+  const realIp = headersList.get('x-real-ip');
+  
+  if (forwardedFor) {
+    return forwardedFor.split(',')[0].trim();
+  }
+  
+  return realIp || 'unknown-ip';
+}
